Introduce linker suggested fixes and run script
When all projects are built with a similar manner that works on
macOS, the Linux linker produces an error because libmymaths cannot
locate libfastmatrix. However, the rpath provided in libmymaths is
correct and if you ldd libmymaths the linker can locate its
dependencies. The linker error can be avoided with Fix1 and Fix2
but that's not an ideal fix as simpleapp is not libfastmatrix
agnostic anymore. The goal is to understand what needs to be done
in order to make the projects work on Linux as they do on macOS,
if possible.